// src/types/common/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.5"
  }
  applicants: {
    Tables: {
      applicant_experiences: {
        Row: {
          company_name: string | null
          description: string
          end_date: string | null
          id: number
          job_title: string | null
          profile_id: string | null
          start_date: string | null
        }
        Insert: {
          company_name?: string | null
          description: string
          end_date?: string | null
          id?: never
          job_title?: string | null
          profile_id?: string | null
          start_date?: string | null
        }
        Update: {
          company_name?: string | null
          description?: string
          end_date?: string | null
          id?: never
          job_title?: string | null
          profile_id?: string | null
          start_date?: string | null
        }
        Relationships: []
      }
      applicant_requirement_media: {
        Row: {
          alt_text: string | null
          applicant_requirement_id: number | null
          created_at: string
          description: string | null
          filename: string | null
          id: number
          mime_type: string | null
          path: string | null
          profiles_id: string | null
          size: number | null
          updated_at: string | null
        }
        Insert: {
          alt_text?: string | null
          applicant_requirement_id?: number | null
          created_at?: string
          description?: string | null
          filename?: string | null
          id?: never
          mime_type?: string | null
          path?: string | null
          profiles_id?: string | null
          size?: number | null
          updated_at?: string | null
        }
        Update: {
          alt_text?: string | null
          applicant_requirement_id?: number | null
          created_at?: string
          description?: string | null
          filename?: string | null
          id?: never
          mime_type?: string | null
          path?: string | null
          profiles_id?: string | null
          size?: number | null
          updated_at?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "applicant_requirement_media_applicant_requirement_id_fkey"
            columns: ["applicant_requirement_id"]
            isOneToOne: false
            referencedRelation: "applicant_requirements"
            referencedColumns: ["id"]
          },
        ]
      }
      applicant_requirements: {
        Row: {
          created_at: string
          id: number
          profile_id: string | null
          remarks: string | null
          requirement_id: number | null
          status: Database["public"]["Enums"]["status_type"] | null
        }
        Insert: {
          created_at?: string
          id?: never
          profile_id?: string | null
          remarks?: string | null
          requirement_id?: number | null
          status?: Database["public"]["Enums"]["status_type"] | null
        }
        Update: {
          created_at?: string
          id?: never
          profile_id?: string | null
          remarks?: string | null
          requirement_id?: number | null
          status?: Database["public"]["Enums"]["status_type"] | null
        }
        Relationships: []
      }
// ...
      applicant_skills: {
        Row: {
          id: number
          profile_id: string | null
          skill_category: string
          skill_name: string
        }
        Insert: {
          id?: never
          profile_id?: string | null
          skill_category: string
          skill_name: string
        }
        Update: {
          id?: never
          profile_id?: string | null
          skill_category?: string
          skill_name?: string
        }
        Relationships: []
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
  core: {
    Tables: {
// ...
      profile_media: {
        Row: {
          alt_text: string | null
          created_at: string
          description: string | null
          filename: string | null
          id: number
          mime_type: string | null
          path: string | null
          profile_id: string | null
          size: number | null
          updated_at: string | null
        }
        Insert: {
          alt_text?: string | null
          created_at?: string
          description?: string | null
          filename?: string | null
          id?: never
          mime_type?: string | null
          path?: string | null
          profile_id?: string | null
          size?: number | null
          updated_at?: string | null
        }
        Update: {
          alt_text?: string | null
          created_at?: string
          description?: string | null
          filename?: string | null
          id?: never
          mime_type?: string | null
          path?: string | null
          profile_id?: string | null
          size?: number | null
          updated_at?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "profile_media_profile_id_fkey"
            columns: ["profile_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      profiles: {
        Row: {
          barangay: string | null
          birthdate: string | null
          contact_number: string | null
          created_at: string
          firstname: string | null
          gender: Database["core"]["Enums"]["gender_type"] | null
          geographic: string | null
          id: string
          is_4ps: boolean | null
          is_pwd: boolean | null
          last_login: string | null
          lastname: string | null
          middlename: string | null
          province: string | null
          region: string | null
          role: Database["core"]["Enums"]["user_role"] | null
          status: Database["core"]["Enums"]["status_type"] | null
          updated_at: string | null
          username: string | null
        }
        Insert: {
          barangay?: string | null
          birthdate?: string | null
          contact_number?: string | null
          created_at?: string
          firstname?: string | null
          gender?: Database["core"]["Enums"]["gender_type"] | null
          geographic?: string | null
          id?: string
          is_4ps?: boolean | null
          is_pwd?: boolean | null
          last_login?: string | null
          lastname?: string | null
          middlename?: string | null
          province?: string | null
          region?: string | null
          role?: Database["core"]["Enums"]["user_role"] | null
          status?: Database["core"]["Enums"]["status_type"] | null
          updated_at?: string | null
          username?: string | null
        }
        Update: {
          barangay?: string | null
          birthdate?: string | null
          contact_number?: string | null
          created_at?: string
          firstname?: string | null
          gender?: Database["core"]["Enums"]["gender_type"] | null
          geographic?: string | null
          id?: string
          is_4ps?: boolean | null
          is_pwd?: boolean | null
          last_login?: string | null
          lastname?: string | null
          middlename?: string | null
          province?: string | null
          region?: string | null
          role?: Database["core"]["Enums"]["user_role"] | null
          status?: Database["core"]["Enums"]["status_type"] | null
          updated_at?: string | null
          username?: string | null
        }
        Relationships: []
      }
    }
// ...
  }
  employers: {
    Tables: {
      companies: {
        Row: {
          business_type: string | null
          company_address: string | null
          company_contact: string | null
          company_description: string | null
          company_email: string | null
          company_name: string | null
          created_at: string
          employee_count: number | null
          id: number
          industry: string | null
          latitude: number | null
          longitude: number | null
          profile_id: string | null
          registration_number: string | null
          updated_at: string | null
          verification_status: Database["public"]["Enums"]["status_type"] | null
          website: string | null
        }
        Insert: {
          business_type?: string | null
          company_address?: string | null
          company_contact?: string | null
          company_description?: string | null
          company_email?: string | null
          company_name?: string | null
          created_at?: string
          employee_count?: number | null
          id?: never
          industry?: string | null
          latitude?: number | null
          longitude?: number | null
          profile_id?: string | null
          registration_number?: string | null
          updated_at?: string | null
          verification_status?:
            | Database["public"]["Enums"]["status_type"]
            | null
          website?: string | null
        }
        Update: {
          business_type?: string | null
          company_address?: string | null
          company_contact?: string | null
          company_description?: string | null
          company_email?: string | null
          company_name?: string | null
          created_at?: string
          employee_count?: number | null
          id?: never
          industry?: string | null
          latitude?: number | null
          longitude?: number | null
          profile_id?: string | null
          registration_number?: string | null
          updated_at?: string | null
          verification_status?:
            | Database["public"]["Enums"]["status_type"]
            | null
          website?: string | null
        }
        Relationships: []
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
  public: {
    Tables: {
// ...
      requirement_templates: {
        Row: {
          created_at: string
          description: string | null
          id: number
          name: string | null
          requirement_type: string | null
          updated_at: string | null
        }
        Insert: {
          created_at?: string
          description?: string | null
          id?: number
          name?: string | null
          requirement_type?: string | null
          updated_at?: string | null
        }
        Update: {
          created_at?: string
          description?: string | null
          id?: number
          name?: string | null
          requirement_type?: string | null
          updated_at?: string | null
        }
        Relationships: []
      }
    }
// ...
  }
// ...
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never